refactor(talk_app): Merge TalkRichObjectFile and TalkRichObjectFilePreview widgets
