+++ packages/neon_framework/packages/talk_app/lib/src/widgets/rich_object/file.dart
@@ -3,8 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:neon_framework/models.dart';
 import 'package:neon_framework/theme.dart';
 import 'package:neon_framework/utils.dart';
+import 'package:neon_framework/widgets.dart';
+import 'package:nextcloud/core.dart' as core;
 import 'package:nextcloud/spreed.dart' as spreed;
-import 'package:talk_app/src/widgets/rich_object/file_preview.dart';
 
 /// Displays a file from a rich object.
 class TalkRichObjectFile extends StatelessWidget {
@@ -21,13 +22,73 @@ class TalkRichObjectFile extends StatelessWidget {
   /// The TextStyle to applied to all text elements in this rich object.
   final TextStyle? textStyle;
 
+  int _parseDimension(({int? $int, String? string}) dimension) => dimension.$int ?? int.parse(dimension.string!);
+
   @override
   Widget build(BuildContext context) {
     Widget child;
 
     if (parameter.previewAvailable == 'yes') {
-      child = TalkRichObjectFilePreview(
-        parameter: parameter,
+      child = LayoutBuilder(
+        builder: (context, constraints) {
+          final devicePixelRatio = MediaQuery.of(context).devicePixelRatio;
+
+          final maxSize = Size(constraints.maxWidth, MediaQuery.sizeOf(context).height / 2);
+
+          Size? logicalSize;
+          var deviceSize = Size(-1, maxSize.height / devicePixelRatio);
+
+          if (parameter.width != null && parameter.height != null) {
+            deviceSize = Size(
+              _parseDimension(parameter.width!).toDouble(),
+              _parseDimension(parameter.height!).toDouble(),
+            );
+
+            // Convert to logical pixels
+            logicalSize = deviceSize / devicePixelRatio;
+
+            // Constrain size to max size but keep aspect ratio
+            if (logicalSize.width > maxSize.width) {
+              logicalSize = logicalSize * (maxSize.width / logicalSize.width);
+            }
+            if (logicalSize.height > maxSize.height) {
+              logicalSize = logicalSize * (maxSize.height / logicalSize.height);
+            }
+
+            // Convert back to device pixels
+            deviceSize = logicalSize * devicePixelRatio;
+          }
+
+          final account = NeonProvider.of<Account>(context);
+          Widget image;
+          if (parameter.mimetype == 'image/gif') {
+            // Previews for animated GIFs are not animated, so we have to request the full file.
+            image = NeonUriImage(
+              account: account,
+              uri: Uri.parse('${account.serverURL}/remote.php/dav/files/${account.username}/${parameter.path!}'),
+            );
+          } else {
+            image = NeonApiImage(
+              account: account,
+              etag: parameter.etag,
+              expires: null,
+              getRequest: (client) => client.core.preview.$getPreviewByFileId_Request(
+                fileId: int.parse(parameter.id),
+                x: deviceSize.width.toInt(),
+                y: deviceSize.height.toInt(),
+                a: core.PreviewGetPreviewByFileIdA.$1,
+              ),
+            );
+          }
+
+          return ConstrainedBox(
+            constraints: logicalSize != null ? BoxConstraints.tight(logicalSize) : BoxConstraints.loose(maxSize),
+            child: Tooltip(
+              message: parameter.name,
+              child: image,
+            ),
+          );
+        },
       );
     } else {
       child = Row(
